+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add missing default og:image [#127](https://github.com/etalab/udata-front/pull/127)
 - Fix 500 on favicon in admin [#126](https://github.com/etalab/udata-front/pull/126)
 - Update search results [#110](https://github.com/etalab/udata-front/pull/110)
+- Fix test error with long reuse title [#133](https://github.com/etalab/udata-front/pull/133) 
 
 ## 2.0.5 (2022-06-14)